refactor phase28 external evidence boundary
